feat(preferences) :wrench: refactor
